requirements.txt
langchain==0.0.154
pyPDF==4.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
openai==0.10.0
